% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eClfull.R
\name{eCl.full}
\alias{eCl.full}
\title{Full Eigenvector Clustering with Optional Replication and Minimum Cluster Size}
\usage{
eCl.full(A, replicate = 0.2, minSize = 1, normalize = FALSE)
}
\arguments{
\item{A}{A symmetric \code{n x n} matrix on which clustering is
to be performed. Assumed to be a type of co-assocation matrix. That is,
higher values in \eqn{A_{ij}} imply that observations \eqn{i} and \eqn{j} are \strong{more} likely
to be in the same cluster.}

\item{replicate}{A numeric value between 0 and 1. This parameter controls
the extent of replication of a selected point in the matrix. If nonzero a random
point is replicated \code{ceiling(replicate*n)} times. Defaults to 0.}

\item{minSize}{An integer specifying the minimum allowed cluster size. Defaults to 1.}

\item{normalize}{A logical value indicating whether to normalize \eqn{\Big(A_{Norm} = D_A^{-1/2}AD_A^{-1/2}\Big)} the matrix
before clustering. Defaults to \code{FALSE}.}
}
\value{
A factor vector of length \code{n} representing the final cluster
assignment of each point after the full clustering process.
}
\description{
This function performs a full eigenvector clustering on a given symmetric (co-association) matrix
with an initial pass followed by a phase focused on separating the initial clusters if needed. It allows for an
optional minimum cluster size. This function builds upon the \code{eCl.one} function.
}
\details{
The function first normalizes the input matrix \code{A} if requested.
It then performs an initial pass of clustering using the \code{\link{eCl.one}}
function with replication turned off. During the breaking phase, the function
refines the clusters by applying \code{\link{eCl.one}} to each unique cluster
with the specified \code{replicate} value. The final cluster assignments are
obtained by combining the results of the initial and breaking phases.
}
\note{
The \code{replicate} value should typically be between 0 and 1.
Setting \code{replicate} to a value greater than or equal to 1 may lead
to unexpected behavior.
}
\examples{
# Create a symmetric matrix
A <- matrix(runif(25), nrow = 5)
A <- (A + t(A)) / 2

# Perform full clustering with default parameters
clust <- eCl.full(A)

# Perform full clustering with normalization and minimum cluster size
clust <- eCl.full(A, normalize = TRUE, minSize = 2)

}
\seealso{
\code{\link{eCl.one}} for the underlying clustering function.
}
