% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eClone.R
\name{eCl.one}
\alias{eCl.one}
\title{Eigenvector Clustering with Optional Replication and Minimum Cluster Size}
\usage{
eCl.one(A, replicate = 0, minSize = 1, normalize = FALSE)
}
\arguments{
\item{A}{A symmetric \code{n x n} matrix on which clustering is
to be performed. Assumed to be a type of co-assocation matrix. That is,
higher values in \eqn{A_{ij}} imply that observations \eqn{i} and \eqn{j} are \strong{more} likely
to be in the same cluster.}

\item{replicate}{A numeric value between 0 and 1. This parameter controls
the extent of replication of a selected point in the matrix. If nonzero a random
point is replicated \code{ceiling(replicate*n)} times. Defaults to 0.}

\item{minSize}{An integer specifying the minimum allowed cluster size. Defaults to 1.}

\item{normalize}{A logical value indicating whether to normalize \eqn{\Big(A_{Norm} = D_A^{-1/2}AD_A^{-1/2}\Big)} the matrix
before clustering. Defaults to \code{FALSE}.}
}
\value{
A numeric vector of length \code{n} representing the cluster
assignment of each point.
}
\description{
This function performs eigenvector clustering on a given symmetric (co-association) matrix.
The function can optionally normalize the input matrix, and replicate a
a single value in the matrix before clustering. The minimum allowable cluster size can
also be specified.
}
\details{
The function first checks that the input matrix \code{A} is square and symmetric.
If the \code{normalize} argument is set to \code{TRUE}, the function normalizes
the matrix by scaling it using the diagonal values. If \code{replicate} is not
zero, a portion of the matrix is replicated. The function then performs an
eigendecomposition of the matrix and uses the resulting eigenvectors and
eigenvalues to cluster the data. If the \code{minSize} argument is greater than 1,
the function ensures that no cluster is smaller than the specified minimum size
by reassigning points from smaller clusters to the next closest option based on the eigendecomposition.
}
\note{
The \code{replicate} value should typically be between 0 and 1.
Setting \code{replicate} to a value greater than or equal to 1 may lead
to unexpected behavior.
}
\examples{
# Create a symmetric matrix
A <- matrix(runif(25), nrow = 5)
A <- (A + t(A)) / 2

# Perform clustering with default parameters
clust <- eCl.one(A)

# Perform clustering with replication and minimum cluster size
clust <- eCl.one(A, replicate = 0.1, minSize = 2)

}
