% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenCluster-package.R
\docType{package}
\name{eigenCluster-package}
\alias{eigenCluster}
\alias{eigenCluster-package}
\title{eigenCluster: An Eigen-Decomposition Method for Identifying a Consensus Clustering from a Co-Association Matrix}
\description{
Implements a consensus clustering algorithm from a co-association matrix based on eigendecomposition. The method does not require the investigator to choose a target number of clusters, and is capable of picking a clustering not identified in the creation of the co-association matrix. Full details to be specified in a forthcoming paper.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukeDuttweiler/eigenCluster}
  \item Report bugs at \url{https://github.com/LukeDuttweiler/eigenCluster/issues}
}

}
\author{
\strong{Maintainer}: Luke Duttweiler \email{lduttweiler@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-0467-995X}{ORCID})

}
\keyword{internal}
