% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFConsensus.R
\name{lf.oneIter}
\alias{lf.oneIter}
\title{Single Iteration of iGraph clustering to support LF Consensus Clustering}
\usage{
lf.oneIter(adjMat, np, threshold, alg, ...)
}
\arguments{
\item{adjMat}{A square, symmetric adjacency matrix representing the network, with edge weights between nodes.}

\item{np}{Integer. The number of partitions to generate in each iteration (default is 20).}

\item{threshold}{Numeric. A threshold for filtering edges; values below this are set to zero (default is 0.2).}

\item{alg}{Character. The community detection algorithm to use, with 'louvain' as the default.
The algorithm should be compatible with the \code{igraph} package, ie. \code{cluster_}alg.}

\item{...}{Additional arguments passed to the community detection algorithm specified in \code{alg}.}
}
\value{
A new adjacency matrix, updated based on averaged partitions.
}
\description{
Helper function to perform a single iteration of consensus clustering on an adjacency matrix
by generating multiple partitions and averaging them into a new adjacency matrix.
}
\details{
The function creates multiple partitions, adjusts singleton nodes to improve connectivity,
and computes an averaged adjacency matrix across partitions.
}
