% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustSpectral.R
\name{robustSpectral}
\alias{robustSpectral}
\title{Robust Spectral Ensemble Clustering}
\usage{
robustSpectral(S, K, lambda1, lambda2)
}
\arguments{
\item{S}{Co-association matrix}

\item{K}{Number of clusters to identify}

\item{lambda1}{Tuning parameter 1. From the paper, works best in (.1, 1)}

\item{lambda2}{Tuning parameter 2, From the paper, works best in (.1, 1)}
}
\value{
Cluster membership of observations, order matches row/column order in S.
}
\description{
This function implements Algorithm 1 from 'Robust Spectral Ensemble Clustering via Rank Minimization' by Tao et al. (2019), with some small adjustments found in the code given on Z. Tao's own website at https://ztao.cc/publications.html.
}
