% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{svThresholdOp}
\alias{svThresholdOp}
\title{svThresholdOp}
\usage{
svThresholdOp(M, tau)
}
\arguments{
\item{M}{A matrix}

\item{tau}{A scalar, used for soft shrinkage of singular values}
}
\value{
Matrix with same dimensions as M, shrunk according to the operator.
}
\description{
Implements the Singular Value Thresholding operator as described in 'A SINGULAR VALUE THRESHOLDING ALGORITHM FORMATRIX COMPLETION', Cai et al. (2010)
}
